fix(di): add global flag to register correctly provider on GlobalRegistry vs injector.container
